sync: community changes

+++ apiserver/plane/app/views/cycle/base.py
@@ -19,7 +19,7 @@ from django.db.models import (
     When,
     Subquery,
     Sum,
-    IntegerField,
+    FloatField,
 )
 from django.db.models.functions import Coalesce, Cast
 from django.utils import timezone
@@ -87,7 +87,7 @@ class CycleViewSet(BaseViewSet):
             .values("issue_cycle__cycle_id")
             .annotate(
                 backlog_estimate_point=Sum(
-                    Cast("estimate_point__value", IntegerField())
+                    Cast("estimate_point__value", FloatField())
                 )
             )
             .values("backlog_estimate_point")[:1]
@@ -101,7 +101,7 @@ class CycleViewSet(BaseViewSet):
             .values("issue_cycle__cycle_id")
             .annotate(
                 unstarted_estimate_point=Sum(
-                    Cast("estimate_point__value", IntegerField())
+                    Cast("estimate_point__value", FloatField())
                 )
             )
             .values("unstarted_estimate_point")[:1]
@@ -115,7 +115,7 @@ class CycleViewSet(BaseViewSet):
             .values("issue_cycle__cycle_id")
             .annotate(
                 started_estimate_point=Sum(
-                    Cast("estimate_point__value", IntegerField())
+                    Cast("estimate_point__value", FloatField())
                 )
             )
             .values("started_estimate_point")[:1]
@@ -129,7 +129,7 @@ class CycleViewSet(BaseViewSet):
             .values("issue_cycle__cycle_id")
             .annotate(
                 cancelled_estimate_point=Sum(
-                    Cast("estimate_point__value", IntegerField())
+                    Cast("estimate_point__value", FloatField())
                 )
             )
             .values("cancelled_estimate_point")[:1]
@@ -143,7 +143,7 @@ class CycleViewSet(BaseViewSet):
             .values("issue_cycle__cycle_id")
             .annotate(
                 completed_estimate_points=Sum(
-                    Cast("estimate_point__value", IntegerField())
+                    Cast("estimate_point__value", FloatField())
                 )
             )
             .values("completed_estimate_points")[:1]
@@ -156,7 +156,7 @@ class CycleViewSet(BaseViewSet):
             .values("issue_cycle__cycle_id")
             .annotate(
                 total_estimate_points=Sum(
-                    Cast("estimate_point__value", IntegerField())
+                    Cast("estimate_point__value", FloatField())
                 )
             )
             .values("total_estimate_points")[:1]
@@ -288,37 +288,37 @@ class CycleViewSet(BaseViewSet):
             .annotate(
                 backlog_estimate_points=Coalesce(
                     Subquery(backlog_estimate_point),
-                    Value(0, output_field=IntegerField()),
+                    Value(0, output_field=FloatField()),
                 ),
             )
             .annotate(
                 unstarted_estimate_points=Coalesce(
                     Subquery(unstarted_estimate_point),
-                    Value(0, output_field=IntegerField()),
+                    Value(0, output_field=FloatField()),
                 ),
             )
             .annotate(
                 started_estimate_points=Coalesce(
                     Subquery(started_estimate_point),
-                    Value(0, output_field=IntegerField()),
+                    Value(0, output_field=FloatField()),
                 ),
             )
             .annotate(
                 cancelled_estimate_points=Coalesce(
                     Subquery(cancelled_estimate_point),
-                    Value(0, output_field=IntegerField()),
+                    Value(0, output_field=FloatField()),
                 ),
             )
             .annotate(
                 completed_estimate_points=Coalesce(
                     Subquery(completed_estimate_point),
-                    Value(0, output_field=IntegerField()),
+                    Value(0, output_field=FloatField()),
                 ),
             )
             .annotate(
                 total_estimate_points=Coalesce(
                     Subquery(total_estimate_point),
-                    Value(0, output_field=IntegerField()),
+                    Value(0, output_field=FloatField()),
                 ),
             )
             .order_by("-is_favorite", "name")
@@ -396,12 +396,12 @@ class CycleViewSet(BaseViewSet):
                         .values("display_name", "assignee_id", "avatar")
                         .annotate(
                             total_estimates=Sum(
-                                Cast("estimate_point__value", IntegerField())
+                                Cast("estimate_point__value", FloatField())
                             )
                         )
                         .annotate(
                             completed_estimates=Sum(
-                                Cast("estimate_point__value", IntegerField()),
+                                Cast("estimate_point__value", FloatField()),
                                 filter=Q(
                                     completed_at__isnull=False,
                                     archived_at__isnull=True,
@@ -411,7 +411,7 @@ class CycleViewSet(BaseViewSet):
                         )
                         .annotate(
                             pending_estimates=Sum(
-                                Cast("estimate_point__value", IntegerField()),
+                                Cast("estimate_point__value", FloatField()),
                                 filter=Q(
                                     completed_at__isnull=True,
                                     archived_at__isnull=True,
@@ -434,12 +434,12 @@ class CycleViewSet(BaseViewSet):
                         .values("label_name", "color", "label_id")
                         .annotate(
                             total_estimates=Sum(
-                                Cast("estimate_point__value", IntegerField())
+                                Cast("estimate_point__value", FloatField())
                             )
                         )
                         .annotate(
                             completed_estimates=Sum(
-                                Cast("estimate_point__value", IntegerField()),
+                                Cast("estimate_point__value", FloatField()),
                                 filter=Q(
                                     completed_at__isnull=False,
                                     archived_at__isnull=True,
@@ -449,7 +449,7 @@ class CycleViewSet(BaseViewSet):
                         )
                         .annotate(
                             pending_estimates=Sum(
-                                Cast("estimate_point__value", IntegerField()),
+                                Cast("estimate_point__value", FloatField()),
                                 filter=Q(
                                     completed_at__isnull=True,
                                     archived_at__isnull=True,
@@ -837,12 +837,12 @@ class CycleViewSet(BaseViewSet):
                 .values("display_name", "assignee_id", "avatar")
                 .annotate(
                     total_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField())
+                        Cast("estimate_point__value", FloatField())
                     )
                 )
                 .annotate(
                     completed_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField()),
+                        Cast("estimate_point__value", FloatField()),
                         filter=Q(
                             completed_at__isnull=False,
                             archived_at__isnull=True,
@@ -852,7 +852,7 @@ class CycleViewSet(BaseViewSet):
                 )
                 .annotate(
                     pending_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField()),
+                        Cast("estimate_point__value", FloatField()),
                         filter=Q(
                             completed_at__isnull=True,
                             archived_at__isnull=True,
@@ -875,12 +875,12 @@ class CycleViewSet(BaseViewSet):
                 .values("label_name", "color", "label_id")
                 .annotate(
                     total_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField())
+                        Cast("estimate_point__value", FloatField())
                     )
                 )
                 .annotate(
                     completed_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField()),
+                        Cast("estimate_point__value", FloatField()),
                         filter=Q(
                             completed_at__isnull=False,
                             archived_at__isnull=True,
@@ -890,7 +890,7 @@ class CycleViewSet(BaseViewSet):
                 )
                 .annotate(
                     pending_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField()),
+                        Cast("estimate_point__value", FloatField()),
                         filter=Q(
                             completed_at__isnull=True,
                             archived_at__isnull=True,
@@ -1235,12 +1235,12 @@ class TransferCycleIssueEndpoint(BaseAPIView):
                 .values("display_name", "assignee_id", "avatar")
                 .annotate(
                     total_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField())
+                        Cast("estimate_point__value", FloatField())
                     )
                 )
                 .annotate(
                     completed_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField()),
+                        Cast("estimate_point__value", FloatField()),
                         filter=Q(
                             completed_at__isnull=False,
                             archived_at__isnull=True,
@@ -1250,7 +1250,7 @@ class TransferCycleIssueEndpoint(BaseAPIView):
                 )
                 .annotate(
                     pending_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField()),
+                        Cast("estimate_point__value", FloatField()),
                         filter=Q(
                             completed_at__isnull=True,
                             archived_at__isnull=True,
@@ -1289,12 +1289,12 @@ class TransferCycleIssueEndpoint(BaseAPIView):
                 .values("label_name", "color", "label_id")
                 .annotate(
                     total_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField())
+                        Cast("estimate_point__value", FloatField())
                     )
                 )
                 .annotate(
                     completed_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField()),
+                        Cast("estimate_point__value", FloatField()),
                         filter=Q(
                             completed_at__isnull=False,
                             archived_at__isnull=True,
@@ -1304,7 +1304,7 @@ class TransferCycleIssueEndpoint(BaseAPIView):
                 )
                 .annotate(
                     pending_estimates=Sum(
-                        Cast("estimate_point__value", IntegerField()),
+                        Cast("estimate_point__value", FloatField()),
                         filter=Q(
                             completed_at__isnull=True,
                             archived_at__isnull=True,
